updated graphql-service template files: models.ts was restored and it imports all models from ./models/index.ts
