+++ packages/cli-scaffold-graphql-service/templateFiles/models/index.ts
@@ -1,36 +1,3 @@
-import { flow } from "lodash";
-import { withStorage, withCrudLogs, withSoftDelete, withFields } from "@webiny/commodo";
-import { withUser } from "@webiny/api-security";
 import myModel from "./myModel.model";
 
-export default () => ({
-    name: "graphql-context-models-...",
-    type: "graphql-context",
-    apply(context) {
-        const driver = context.commodo && context.commodo.driver;
-
-        if (!driver) {
-            throw Error(
-                `Commodo driver is not configured! Make sure you add a Commodo driver plugin to your service.`
-            );
-        }
-
-        const createBase = () =>
-            flow(
-                withFields({
-                    id: context.commodo.fields.id()
-                }),
-                withStorage({ driver }),
-                withUser(context),
-                withSoftDelete(),
-                withCrudLogs()
-            )();
-
-        const MyModel = myModel({ createBase });
-
-        context.models = {
-            MyModel,
-            createBase
-        };
-    }
-});
+export default { myModel };
